Bugfix on TryToElementsIndex when key is String and key is larger than int32 maximum
Issue: https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I7VDGH

+++ ecmascript/ic/ic_runtime_stub.h
@@ -71,7 +71,7 @@ public:
     static inline JSTaggedValue LoadElement(JSObject *receiver, JSTaggedValue key);
     static inline JSTaggedValue StoreElement(JSThread *thread, JSObject *receiver, JSTaggedValue key,
                                              JSTaggedValue value, JSTaggedValue handlerInfo);
-    static inline int32_t TryToElementsIndex(JSTaggedValue key);
+    static inline int64_t TryToElementsIndex(JSTaggedValue key);
     static inline JSTaggedValue LoadMiss(JSThread *thread, ProfileTypeInfo *profileTypeInfo, JSTaggedValue receiver,
                                          JSTaggedValue key, uint32_t slotId, ICKind kind);
     static inline JSTaggedValue StoreMiss(JSThread *thread, ProfileTypeInfo *profileTypeInfo, JSTaggedValue receiver,
